requirements.txt
# Núcleo
numpy<2.0
pandas

# Processamento de vídeo
opencv-python
Pillow
scenedetect[opencv] # para detecção de cena

# Deep Learning / Embeddings
torch
torchvision
ftfy
regex
tqdm

# CLIP (OpenAI) - opcional, mas recomendado para embeddings semânticos
# pode exigir git instalado
git+https://github.com/openai/CLIP.git

# Áudio (opcional - multimodal)
librosa
soundfile

# Elasticsearch (cliente Python)
elasticsearch==8.12.0

# Utilidades
scikit-learn

# Dataset
fsspec<=2025.10.0 # para evitar conflitos
yt-dlp
